+++ lib/headless.js
@@ -10,47 +10,7 @@ const { Vec3 } = require('vec3')
 const { WorldRenderer } = require('../src/worldrenderer')
 const { makeTextureAtlas } = require('../lib/atlas')
 const { prepareBlocksStates } = require('../lib/models')
-
-function getBufferFromStream (stream) {
-  return new Promise(
-    (resolve, reject) => {
-      let buffer = Buffer.from([])
-      stream.on('data', buf => {
-        buffer = Buffer.concat([buffer, buf])
-      })
-      stream.on('end', () => resolve(buffer))
-      stream.on('error', reject)
-    }
-  )
-}
-
-function spiral (X, Y, fun) { // TODO: move that to spiralloop package
-  let x = 0
-  let y = 0
-  let dx = 0
-  let dy = -1
-  const N = Math.max(X, Y) * Math.max(X, Y)
-  const hX = X / 2
-  const hY = Y / 2
-  for (let i = 0; i < N; i++) {
-    if (-hX < x && x <= hX && -hY < y && y <= hY) {
-      fun(x, y)
-    }
-    if (x === y || (x < 0 && x === -y) || (x > 0 && x === 1 - y)) {
-      const tmp = dx
-      dx = -dy
-      dy = tmp
-    }
-    x += dx
-    y += dy
-  }
-}
-
-function chunkPos (pos) {
-  const x = Math.floor(pos.x / 16)
-  const z = Math.floor(pos.z / 16)
-  return [x, z]
-}
+const { getBufferFromStream, spiral, chunkPos } = require('./utils')
 
 module.exports = (bot, { viewDistance = 6, output = 'output.mp4', frames = 200, width = 512, height = 512 }) => {
   const mcAssets = require('minecraft-assets')(bot.version)
